Fixed various syntax issues

+++ cpp/bmi.h
@@ -59,7 +59,7 @@ class Bmi
     virtual int get_grid_point_count(int id, int* dest) const = 0;
     virtual int get_grid_vertex_count(int id, int* dest) const = 0;
     virtual int get_grid_connectivity(int id, int* dest) const = 0;
-    virtual int get_grid_offset(int id, int* dest) const = 0;
+    virtual int get_grid_offset(int id, double* dest) const = 0;
 };
 
 // TODO: converter to C-BMI struct
